nfs: Add '--with-nss-modules' configure option to specify nss modules
From: Sami Wagiaalla <swagiaal@redhat.com>

I was having trouble setting up NFS on Atomic Host.  It turns out
there is an issue when rpcbind is trying to find the uid of the
rpc user.  OSTree based operating systems store user information
for system users such as the rpc user in /usr/lib/passwd and
leaves /etc/passwd for humans users.  This is enabled by the use
of the nss module nss-altfiles which allows one to specify
additional files to be added the the passwd database.  rpcbind
however overrides the rule added to /etc/nsswitch.conf and removes
"altfiles" from the list of modules by doing the following:

    __nss_configure_lookup("passwd", "files");

This was added in commit 77f7556878d1fe03dc ("[...]use
__nss_configure_lookup() to restrict the [rpc user] lookup") to
remove "nis" form the list of modules and prevent rpcbind from
having a circular dependency on itself.  In an OSTree based
operating system however this prevents rpcbind from finding the rpc
user and the service cannot start.

This patch adds an option --with-nss-modules which allows one
to specify the nss modules which should be searched for user
information.  The default setting is "files" which preserves the
current behavior, but this enables one to add other modules to
the search path.

Signed-off-by: Sami Wagiaalla <swagiaal@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>